+++ ffi/src/expressions/mod.rs
@@ -3,8 +3,8 @@
 use delta_kernel::Expression;
 use delta_kernel_ffi_macros::handle_descriptor;
 
-pub mod engine_expr_visitor;
-pub mod kernel_expr_visitor;
+pub mod engine;
+pub mod kernel;
 
 #[handle_descriptor(target=Expression, mutable=false, sized=true)]
 pub struct SharedExpression;
